+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - feat: use Grafana 12.1.0 (#4855 - @swiffer)
 - feat(nix): use datasources.settings.datasources to allow merging Grafana sources from multiple modules (#4870 - @JakobLichterfeld)
 - fix(nix): correctly set default Grafana dashboard path (#4870 - @JakobLichterfeld)
+- feat(nix): allow disabling default Grafana dashboard with setDefaultDashboard option (#4870 - @JakobLichterfeld)
 
 #### Build, CI, internal